Add ability to specify the server address used by Configuration Data Packages.
This commit lets a server administrator to set a specific address to
use as the server address in a client Configuration Data Package.

By default, this value is empty, and the behavior will be unchanged:
the address used is the hostname of the API request sender's.
If the value is not empty, it is used instead.